backend/requirements.txt
Django==6.0.2
django-cors-headers==4.0.0
PyPDF2==3.0.0
python-docx==0.8.11
